Add production Dockerfile.
I had to make a small change to the ncurses initialization. Running it out of
this Docker image ends up screwing up the shell but it can be fixed with
`reset`. I think what's happening is that I need to respond to the signal and
clean up ncurses. I'm not sure how to do that at the moment.